move redis directory
